Next2

+++ benchmark/src/main/scala/benchmark/bench/StatusLog.scala
@@ -1,8 +1,8 @@
 package benchmark.bench
 
-import fabric.rw.RW
+import fabric.rw._
 
-case class StatusLog(progress: Double,
+case class StatusLog(progress: Long,
                      timeStamp: Long,
                      elapsed: Double,
                      heap: Long,
